requirements.txt

strands-agents>=1.7.0
GitPython>=3.1.43
PyYAML>=6.0.2
pydantic>=2.8.2
requests>=2.32.3
docker>=7.1.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
